wp-cron operational for JS usability
WP cron was setup so we aren't hitting the YouTube API every time
someone loads the page. The cron is setup for 30 seconds. When we're
hitting wp-cron.php with JS requests it will create our JSON file...
which we will pull with JS a moment later.... WE ARE READY FOR JS
NOW!!!