JS implemented along with cleanup
The JS has been made to work when the channel goes live. Also some
cleanup has been done. We are still using methods that could probably be
combined into one but that is for a future release. This needs to get
done. It will be pushed to master once testing has been done on the
previous commits.